Move snap() to math directory
