+++ packages/perseus/src/widgets/interactive-graphs/graphs/polygon.tsx
@@ -2,7 +2,8 @@ import {Polygon, vec} from "mafs";
 import * as React from "react";
 
 import {moveAll, movePoint} from "../reducer/interactive-graph-action";
-import {TARGET_SIZE, snap} from "../utils";
+import {TARGET_SIZE} from "../utils";
+import {snap} from "../math"
 
 import {Angle} from "./components/angle";
 import {StyledMovablePoint} from "./components/movable-point";
