+++ packages/perseus/src/widgets/interactive-graphs/math/clamp.ts
@@ -1,0 +1,10 @@
+// Restricts `value` to be between `min` and `max`.
+export function clamp(value: number, min: number, max: number): number {
+    if (value < min) {
+        return min;
+    }
+    if (value > max) {
+        return max;
+    }
+    return value;
+}
